structural changes:  - split code among 'include' and 'src' folder  - removed local TSVDUnfold implementation (always use ROOT)  - always compile TUnfold wrapper
