+++ include/RooUnfold_LinkDef.h
@@ -14,14 +14,9 @@
 #pragma link C++ class RooUnfoldErrors+;
 #pragma link C++ class RooUnfoldParms+;
 #pragma link C++ class RooUnfoldInvert+;
-#ifndef NOTUNFOLD
 #pragma link C++ class RooUnfoldTUnfold+;
-#endif
 #ifdef HAVE_DAGOSTINI
 #pragma link C++ class RooUnfoldDagostini+;
-#endif
-#ifdef HAVE_TSVDUNFOLD
-#pragma link C++ class TSVDUnfold+;
 #endif
 
 #endif
