+++ src/RooUnfold.cxx
@@ -62,7 +62,7 @@ END_HTML */
 
 /////////////////////////////////////////////////////////////
 
-#include "RooUnfold.h"
+#include "../include/RooUnfold.h"
 
 #include <iostream>
 #include <iomanip>
@@ -81,18 +81,18 @@ END_HTML */
 #include "TDecompChol.h"
 #include "TRandom.h"
 
-#include "RooUnfoldResponse.h"
-#include "RooUnfoldErrors.h"
+#include "../include/RooUnfoldResponse.h"
+#include "../include/RooUnfoldErrors.h"
 // Need subclasses just for RooUnfold::New()
-#include "RooUnfoldBayes.h"
-#include "RooUnfoldSvd.h"
-#include "RooUnfoldInvert.h"
-#include "RooUnfoldBinByBin.h"
+#include "../include/RooUnfoldBayes.h"
+#include "../include/RooUnfoldSvd.h"
+#include "../include/RooUnfoldInvert.h"
+#include "../include/RooUnfoldBinByBin.h"
 #ifndef NOTUNFOLD
-#include "RooUnfoldTUnfold.h"
+#include "../include/RooUnfoldTUnfold.h"
 #endif
 #ifdef HAVE_DAGOSTINI
-#include "RooUnfoldDagostini.h"
+#include "../include/RooUnfoldDagostini.h"
 #endif
 
 using std::vector;
